01_repaso_sql/requirements.txt
mysql-connector-python==9.1.0
tqdm
